Feature/phu/fpoly showrooms/test demo

+++ app/Http/Controllers/DirectoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Level;
 use App\Models\Project;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class DirectoryController extends Controller
@@ -14,11 +15,16 @@ class DirectoryController extends Controller
      */
     public function index()
     {
-        $directorys = Level::query()->get();
-        return view("directorys.list", compact('directorys'));
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            $directorys = Level::query()->get();
+            return view("directorys.list", compact('directorys'));
+        } else {
+            return redirect()->route('login');
+        }
     }
     public function directory()
     {
+
         $directorys = Level::query()
             ->whereExists(function ($query) {
                 $query->select(DB::raw(1))
@@ -38,7 +44,11 @@ class DirectoryController extends Controller
      */
     public function create(Level $level)
     {
-        return view("directorys.create", compact('level'));
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            return view("directorys.create", compact('level'));
+        } else {
+            return redirect()->route('login');
+        }
     }
 
     /**
@@ -46,50 +56,51 @@ class DirectoryController extends Controller
      */
     public function store(Request $request)
     {
-        $requestData = $request->all();
-        Level::query()->create($requestData);
-        return redirect()->route('directory.index');
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            $requestData = $request->all();
+            Level::query()->create($requestData);
+            return redirect()->route('directory.index');
+        } else {
+            return redirect()->route('login');
+        }
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Level $level)
-    {
-        //
-    }
-
     /**
      * Show the form for editing the specified resource.
      */
     public function edit($level)
     {
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            $directorys = Level::query()->select()->where('id', $level)->get();
 
-        $directorys = Level::query()->select()->where('id', $level)->get();
-
-        return view("directorys.edit", compact('directorys', 'level'));
+            return view("directorys.edit", compact('directorys', 'level'));
+        } else {
+            return redirect()->route('login');
+        }
     }
-    public function details(Level $level)
-    {
-        return view("directorys.details");
-    }
-
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, Level $level)
     {
-        $requestData = $request->all();
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            $requestData = $request->all();
 
-        // DB::enableQueryLog();
-        Level::query()->where('id', $requestData['id'])->update([
-            'name' => $requestData['name'],
-            'description' => $requestData['description'],
-            'updated_at' => $requestData['updated_at'],
-        ]);
-        // $queries = DB::getQueryLog();
-        // dd($queries);
-        return redirect()->route('directory.index');
+            // DB::enableQueryLog();
+            Level::query()->where('id', $requestData['id'])->update([
+                'name' => $requestData['name'],
+                'description' => $requestData['description'],
+                'updated_at' => $requestData['updated_at'],
+            ]);
+            // $queries = DB::getQueryLog();
+            // dd($queries);
+            return redirect()->route('directory.index');
+        } else {
+            return redirect()->route('login');
+        }
     }
 
     /**
@@ -97,14 +108,18 @@ class DirectoryController extends Controller
      */
     public function destroy($level)
     {
-        $count = Project::query()
-            ->where('id', $level)
-            ->count();
-        if ($count > 0) {
-            return redirect()->route('directory.index');
+        if (Auth::check() && Auth::user()->role == 'admin') {
+            $count = Project::query()
+                ->where('id', $level)
+                ->count();
+            if ($count > 0) {
+                return redirect()->route('directory.index');
+            } else {
+                Level::query()->where('id', $level)->delete();
+                return redirect()->route('directory.index');
+            }
         } else {
-            Level::query()->where('id', $level)->delete();
-            return redirect()->route('directory.index');
+            return redirect()->route('login');
         }
     }
 }
